diag(ble): make upMillis the real 64-bit twin on Kotlin
Found while checking the epitaph for hot-path cost. Swift's Int is 64-bit
and Kotlin's is 32-bit, so `upMillis: Int` made the Kotlin formatter the
NARROWER half of a pair the parity contract says are twins.

The consequence is silent and backwards: `(now - since).toInt()` wraps past
~24.8 days, the formatter's own clamp then turns the negative into 0, and a
link held up longer than any other would print "up 0ms" - reading as dead
when it was the healthiest link we ever had. In a diagnostic whose whole job
is telling a live link from a dead one, that is the wrong failure.

Take Long, pass the subtraction straight through, drop the lossy toInt.

No perf change: this is one field on a per-disconnect line. The per-frame
cost stays two integer increments and one UUID compare, against a handler
that already copies the payload into a fresh array and reads the clock.

Kotlin: 22/22, forced re-run.

+++ android/app/src/main/java/com/noop/ble/WhoopBleClient.kt
@@ -9811,7 +9811,7 @@ class WhoopBleClient(
         // hazard the hold-time snapshot below already guards.
         linkUpSinceMs?.let { since ->
             log(ConnectionReadout.linkEpitaph(
-                upMillis = (System.currentTimeMillis() - since).toInt(),
+                upMillis = System.currentTimeMillis() - since,
                 inboundFrames = inboundFrames, inboundBytes = inboundBytes,
                 cmdChannelFrames = cmdChannelFrames, realtimeArmed = realtimeArmedThisLink,
                 ended = "status=$status",
